Feature/pdf history

+++ src/Controller/GeneratePdfController.php
@@ -2,8 +2,13 @@
 
 namespace App\Controller;
 
+use App\Repository\PdfRepository;
 use App\Service\GotenbergService;
+use App\Service\PdfCheckerService;
+use App\Service\PdfFileNameGeneratorService;
+use App\Service\PdfHistoryService;
 use Doctrine\DBAL\Types\Types;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
@@ -13,12 +18,27 @@ use Symfony\Component\Routing\Attribute\Route;
 
 class GeneratePdfController extends AbstractController
 {
-    private $pdfService;
+    private $gotenbergService;
+    private $pdfCheckerService;
+    private $pdfHistoryService;
+    private $pdfFileNameGeneratorService;
+
     private string $publicTempAbsoluteDirectory;
 
-    public function __construct(GotenbergService $pdfService, string $publicTempAbsoluteDirectory)
+    public function __construct(
+        GotenbergService $gotenbergService,
+        PdfCheckerService $pdfCheckerService,
+        PdfRepository $pdfRepository,
+        PdfHistoryService $pdfHistoryService,
+        PdfFileNameGeneratorService $pdfFileNameGeneratorService,
+        string $publicTempAbsoluteDirectory
+    )
     {
-        $this->pdfService = $pdfService;
+        $this->gotenbergService = $gotenbergService;
+        $this->pdfCheckerService = $pdfCheckerService;
+        $this->pdfHistoryService = $pdfHistoryService;
+        $this->pdfFileNameGeneratorService = $pdfFileNameGeneratorService;
+        $this->pdfRepository = $pdfRepository;
         $this->publicTempAbsoluteDirectory = $publicTempAbsoluteDirectory;
 
     }
@@ -32,7 +52,7 @@ class GeneratePdfController extends AbstractController
     }
 
     #[Route('/pdf/generate/url', name: 'app_pdf_generate_url')]
-    public function generatePdfUrl(ParameterBagInterface $parameterBag, Request $request): Response
+    public function generatePdfUrl(ParameterBagInterface $parameterBag, Request $request, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createFormBuilder()
             ->add('url', null, ['required' => true])
@@ -43,11 +63,24 @@ class GeneratePdfController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $url = $form->getData()['url'];
 
-            $pdf = $this->pdfService->generatePdfUrl($parameterBag, $url);
+            //Check si l'user à atteint son quota de génération
+            if ($this->pdfCheckerService->hasRichQuota($this->getUser())) {
+                $this->addFlash('danger', 'Vous avez atteint votre quota de génération de PDFs pour aujourd\'hui (soit '. $this->getUser()->getSubscription()->getPdfLimit() .'/jour).');
+                return $this->redirectToRoute('app_pdf_generate_url');
+            }
 
-            return new Response($pdf, 200, [
+            //Générer le pdf
+            $pdf_file = $this->gotenbergService->generatePdfUrl($parameterBag, $url);
+            $filename = $this->pdfFileNameGeneratorService->generateFileName();
+
+            //Enregistrement dans la table pdf pour l'historique
+            if ($pdf_file != null) {
+                $this->pdfHistoryService->savePdf($filename, "url");
+            }
+
+            return new Response($pdf_file, 200, [
                 'Content-Type' => 'application/pdf',
-                'Content-Disposition' => 'attachment; filename="my-generated-pdf.pdf"',
+                'Content-Disposition' => 'attachment; filename="'.$filename.'"',
             ]);
         }
 
@@ -72,18 +105,27 @@ class GeneratePdfController extends AbstractController
             mkdir($tempDir, 0777, true);
 
             $file->move($tempDir, 'index.html');
-
             $filePath = $tempDir . '/' . 'index.html';
             chmod($filePath, 0777);
 
-            $pdf = $this->pdfService->generatePdfHtml($parameterBag, $filePath);
+            if ($this->pdfCheckerService->hasRichQuota($this->getUser())) {
+                $this->addFlash('danger', 'Vous avez atteint votre quota de génération de PDFs pour aujourd\'hui (soit '. $this->getUser()->getSubscription()->getPdfLimit() .'/jour).');
+                return $this->redirectToRoute('app_pdf_generate_html');
+            }
+
+            $pdf_file = $this->gotenbergService->generatePdfHtml($parameterBag, $filePath);
+            $filename = $this->pdfFileNameGeneratorService->generateFileName();
+
+            if ($pdf_file != null) {
+                $this->pdfHistoryService->savePdf($filename, "html");
+            }
 
             unlink($filePath);
             rmdir($tempDir);
 
-            return new Response($pdf, 200, [
+            return new Response($pdf_file, 200, [
                 'Content-Type' => 'application/pdf',
-                'Content-Disposition' => 'attachment; filename="my-generated-pdf.pdf"',
+                'Content-Disposition' => 'attachment; filename="'.$filename .'"',
             ]);
         }
